reshuffle light code to reduce redundacy
