+++ src/ray_light_calc_b.c
@@ -6,7 +6,7 @@
 /*   By: bhowe <bhowe@student.42singapore.sg>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/09 13:58:49 by bhowe             #+#    #+#             */
-/*   Updated: 2024/10/11 23:54:31 by bhowe            ###   ########.fr       */
+/*   Updated: 2024/10/14 12:01:23 by bhowe            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,42 +52,4 @@ void	calc_color(t_data *data, t_rayparams *rp)
 			rp->diffuse_fin = add_shadow(rp);
 	}
 	rp->color_fin = rgb_get(rgb_add(rp->amb_fin, rp->diffuse_fin));
-}
-
-float	calculate_lighting(t_rayparams *rp, t_light *light)
-{
-	t_vec	lv;
-	float	intensity;
-
-	lv = vector_normalize(vector_subtract(light->position, rp->t_hitpoint));
-	if (vector_dotproduct(lv, rp->t_normal) > 0 && rp->t_norm_flip)
-		rp->t_normal = vector_multiply(rp->t_normal, -1);
-	intensity = vector_dotproduct(lv, rp->t_normal);
-	if (intensity <= EPSILON)
-		intensity = 0;
-	return (intensity * light->brightness);
-}
-
-bool	in_shadow(t_data *data, t_rayparams *rp, t_light *light)
-{
-	int			i;
-	t_rayparams	sp;
-	t_ray		sr;
-
-	sr.origin = vector_add(rp->t_hitpoint, rp->t_normal);
-	sr.vector = vector_normalize(vector_subtract(light->position,
-				rp->t_hitpoint));
-	rp->dl = vector_length(vector_subtract(sr.origin, light->position));
-	i = -1;
-	while (++i < data->prim_count)
-	{
-		if (i == rp->prim_num)
-			continue ;
-		if (hit_prim(&sr, data->prims[i], &sp))
-		{
-			if (sp.t > EPSILON && sp.t < rp->dl)
-				return (true);
-		}
-	}
-	return (false);
 }
